Moved tests for package consistency
